Updated library page to include a list of all the books currently inside that library.
